moved the module class registration to the main.js again (should be outside the core app)
